+++ model/subscription.go
@@ -764,7 +764,7 @@ func PurchaseSubscriptionWithBalance(userId int, planId int) error {
 		if err := lockForUpdate(tx).Where("id = ?", userId).First(&user).Error; err != nil {
 			return err
 		}
-		if requiredQuota > 0 && user.Quota < requiredQuota {
+		if requiredQuota > 0 && user.Quota < int64(requiredQuota) {
 			return errors.New("余额不足")
 		}
 		if requiredQuota > 0 {
